+++ src/latex-commands/styling.ts
@@ -321,7 +321,7 @@ defineFunction('selectfont', '', {
 
 // \bf works in any mode
 // As per the LaTeX 2.09 semantics, it overrides shape, family
-defineFunction('bf', '{:rest}', {
+defineFunction('bf', '{:rest*}', {
   applyStyle: () => ({ fontSeries: 'b', fontShape: 'n', fontFamily: 'roman' }),
 });
 
@@ -344,24 +344,24 @@ defineFunction('bold', '{:math*}', {
   applyStyle: () => ({ variantStyle: 'bold' }),
 });
 
-defineFunction('bfseries', '{:rest}', {
+defineFunction('bfseries', '{:rest*}', {
   applyMode: 'text',
   applyStyle: () => ({ fontSeries: 'b' }),
 });
-defineFunction('mdseries', '{:rest}', {
+defineFunction('mdseries', '{:rest*}', {
   applyMode: 'text',
   applyStyle: () => ({ fontSeries: 'm' }),
 });
-defineFunction('upshape', '{:rest}', {
+defineFunction('upshape', '{:rest*}', {
   applyMode: 'text',
   applyStyle: () => ({ fontShape: 'n' }),
 });
-defineFunction('slshape', '{:rest}', {
+defineFunction('slshape', '{:rest*}', {
   applyMode: 'text',
   applyStyle: () => ({ fontShape: 'sl' }),
 });
 // Small caps
-defineFunction('scshape', '{:rest}', {
+defineFunction('scshape', '{:rest*}', {
   applyMode: 'text',
   applyStyle: () => ({ fontShape: 'sc' }),
 });
@@ -454,7 +454,7 @@ defineFunction('mathtt', '{:math*}', {
   applyStyle: () => ({ variant: 'monospace', variantStyle: 'up' }),
 });
 
-defineFunction('it', '{:rest}', {
+defineFunction('it', '{:rest*}', {
   applyStyle: () => ({
     fontSeries: 'm',
     fontShape: 'it',
@@ -464,15 +464,15 @@ defineFunction('it', '{:rest}', {
 });
 
 // In LaTeX, \rmfamily, \sffamily and \ttfamily are no-op in math mode.
-defineFunction('rmfamily', '', {
+defineFunction('rmfamily', '{:rest*}', {
   applyStyle: () => ({ fontFamily: 'roman' }),
 });
 
-defineFunction('sffamily', '', {
+defineFunction('sffamily', '{:rest*}', {
   applyStyle: () => ({ fontFamily: 'sans-serif' }),
 });
 
-defineFunction('ttfamily', '', {
+defineFunction('ttfamily', '{:rest*}', {
   applyStyle: () => ({ fontFamily: 'monospace' }),
 });